separate some functions and create new classes for that
