--- task_12/src/components/home/index.pug
@@ -1,133 +1,0 @@
-doctype html
-html(lang="en")
-head
-    link(rel="stylesheet" href="//use.fontawesome.com/releases/v5.0.4/css/all.css")
-    title Pizza order
-body
-    .container
-        .header
-            .top-time
-                time 12:12:12
-            .top-logo 
-                img(src=require('./images/captainamerica.png'))
-            .top-right-info
-                a.login(href="#") login
-                    //- i.fas.fa-user(aria-hidden="true") &nbsp;&nbsp;login
-                     
-        main.main
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-            .pizza-container
-                .pizza-img
-                    img.img(src=require('./images/pizza-PomoLivo.png'))
-                .pizza-order-info
-                    .div1
-                        span xx
-                        span xx
-                    .div2
-                        span xx
-                        span xx
-
-        footer.footer
-            .left-info
-                p Kottans, Kottans Str.1 tel: 577-788-88
-            .right-info
-                p Pizza Manager © 2018
