--- task_12/src/components/login/login.scss
@@ -1,33 +1,0 @@
-.registration-form{      
-        width: 20%;
-        margin: auto;
-        background: $black;
-        padding: rem(75px) rem(100px);
-        text-align: center;
-
-        & input {
-            display: block;
-            padding: rem(10px) rem(15px);
-            margin: rem(15px) auto;            outline: 1;
-            box-sizing: border-box;
-            width: 100%;
-        }
-    }
-
-    .registration-form__title {
-        color: $white;
-        text-transform: uppercase;
-    }
-
-    .btn-registration {
-        padding: rem(10px) 0;
-        color: $white;
-        border: rem(8px) solid green;
-        text-transform: uppercase;
-        background: green;
-        width: 100%;
-    }
-
-    .message {
-        color: $grey;
-    }
